extract emoji reaction rendering into its own component
