+++ lib/views/issue-detail-view.js
@@ -7,18 +7,10 @@ import IssueTimelineController from '../controllers/issue-timeline-controller';
 import Octicon from '../atom/octicon';
 import IssueishBadge from '../views/issueish-badge';
 import GithubDotcomMarkdown from '../views/github-dotcom-markdown';
+import EmojiReactionsView from '../views/emoji-reactions-view';
 import PeriodicRefresher from '../periodic-refresher';
 import {autobind} from '../helpers';
 import {addEvent} from '../reporter-proxy';
-
-const reactionTypeToEmoji = {
-  THUMBS_UP: '👍',
-  THUMBS_DOWN: '👎',
-  LAUGH: '😆',
-  HOORAY: '🎉',
-  CONFUSED: '😕',
-  HEART: '❤️',
-};
 
 export class BareIssueDetailView extends React.Component {
   static propTypes = {
@@ -65,7 +57,7 @@ export class BareIssueDetailView extends React.Component {
 
   constructor(props) {
     super(props);
-    autobind(this, 'handleRefreshClick', 'refresh', 'renderEmojiReactions', 'renderIssueBody');
+    autobind(this, 'handleRefreshClick', 'refresh', 'renderIssueBody');
   }
 
   componentDidMount() {
@@ -90,27 +82,12 @@ export class BareIssueDetailView extends React.Component {
           html={issueish.bodyHTML || '<em>No description provided.</em>'}
           switchToIssueish={this.props.switchToIssueish}
         />
-        {this.renderEmojiReactions(issueish)}
+        <EmojiReactionsView reactionGroups={issueish.reactionGroups} />
         <IssueTimelineController
           {...childProps}
           switchToIssueish={this.props.switchToIssueish}
         />
       </Fragment>
-    );
-  }
-
-  renderEmojiReactions(issueish) {
-    return (
-      <div className="github-IssueishDetailView-reactions">
-        {issueish.reactionGroups.map(group => (
-          group.users.totalCount > 0
-            ? <span className={cx('github-IssueishDetailView-reactionsGroup', group.content.toLowerCase())}
-              key={group.content}>
-              {reactionTypeToEmoji[group.content]} &nbsp; {group.users.totalCount}
-            </span>
-            : null
-        ))}
-      </div>
     );
   }
