+++ lib/views/pr-detail-view.js
@@ -9,20 +9,12 @@ import PrStatusesView from '../views/pr-statuses-view';
 import Octicon from '../atom/octicon';
 import IssueishBadge from '../views/issueish-badge';
 import GithubDotcomMarkdown from '../views/github-dotcom-markdown';
+import EmojiReactionsView from '../views/emoji-reactions-view';
 import PeriodicRefresher from '../periodic-refresher';
 import {EnableableOperationPropType} from '../prop-types';
 import {autobind} from '../helpers';
 import {addEvent} from '../reporter-proxy';
 import PrCommitsView from '../views/pr-commits-view';
-
-const reactionTypeToEmoji = {
-  THUMBS_UP: '👍',
-  THUMBS_DOWN: '👎',
-  LAUGH: '😆',
-  HOORAY: '🎉',
-  CONFUSED: '😕',
-  HEART: '❤️',
-};
 
 function createCheckoutState(name) {
   return function(cases) {
@@ -88,7 +80,7 @@ export class BarePullRequestDetailView extends React.Component {
 
   constructor(props) {
     super(props);
-    autobind(this, 'handleRefreshClick', 'refresh', 'renderEmojiReactions', 'renderPullRequestBody');
+    autobind(this, 'handleRefreshClick', 'refresh', 'renderPullRequestBody');
   }
 
   componentDidMount() {
@@ -150,7 +142,7 @@ export class BarePullRequestDetailView extends React.Component {
             html={issueish.bodyHTML || '<em>No description provided.</em>'}
             switchToIssueish={this.props.switchToIssueish}
           />
-          {this.renderEmojiReactions(issueish)}
+          <EmojiReactionsView reactionGroups={issueish.reactionGroups} />
           <PrTimelineContainer
             {...childProps}
             switchToIssueish={this.props.switchToIssueish}
@@ -170,21 +162,6 @@ export class BarePullRequestDetailView extends React.Component {
           <PrCommitsView pullRequest={issueish} />
         </TabPanel>
       </Tabs>
-    );
-  }
-
-  renderEmojiReactions(issueish) {
-    return (
-      <div className="github-IssueishDetailView-reactions">
-        {issueish.reactionGroups.map(group => (
-          group.users.totalCount > 0
-            ? <span className={cx('github-IssueishDetailView-reactionsGroup', group.content.toLowerCase())}
-              key={group.content}>
-              {reactionTypeToEmoji[group.content]} &nbsp; {group.users.totalCount}
-            </span>
-            : null
-        ))}
-      </div>
     );
   }
